89 bug enemy attack not disrupted (#100)
* Fixed attack not being interrupted

The reason some attacks weren't getting interrupted properly was because we were sending the backoff signal during a transition between the idle and attack states. The animator didn't "know" which state to use as its source of interruption. We have to specify it in the transition.

* Specified transition interruption sources for all transitions

This is to make sure we don't have any delayed transitions